+++ lib/src/model/i18n_config.dart
@@ -1,4 +1,5 @@
 import 'package:fast_i18n/src/model/build_config.dart';
+import 'package:fast_i18n/src/model/context_type.dart';
 import 'package:fast_i18n/src/model/i18n_locale.dart';
 import 'package:fast_i18n/src/model/pluralization.dart';
 
@@ -14,6 +15,7 @@ class I18nConfig {
   final String enumName;
   final TranslationClassVisibility translationClassVisibility;
   final bool renderFlatMap;
+  final List<ContextType> contexts;
 
   I18nConfig({
     required this.nullSafety,
@@ -26,5 +28,6 @@ class I18nConfig {
     required this.enumName,
     required this.translationClassVisibility,
     required this.renderFlatMap,
+    required this.contexts,
   });
 }